+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [Unreleased]
+### Added
+- documentation to extend `Release` class + test coverage
+- ability to use extended `Release` within `parser`
 
 ## [0.9.1] - 2019-11-26
 ### Fixed
@@ -138,6 +143,7 @@ First version
 [#9]: https://github.com/oscarotero/keep-a-changelog/issues/9
 [#10]: https://github.com/oscarotero/keep-a-changelog/issues/10
 
+[Unreleased]: https://github.com/oscarotero/keep-a-changelog/compare/v0.9.1...HEAD
 [0.9.1]: https://github.com/oscarotero/keep-a-changelog/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.8.2...v0.9.0
 [0.8.2]: https://github.com/oscarotero/keep-a-changelog/compare/v0.8.1...v0.8.2